% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ai.create_image_grok}
\alias{ai.create_image_grok}
\title{Generate image with Grok (which uses Flux).
Note: limitation of the prompt is about 1000 characters.}
\usage{
ai.create_image_grok(
  prompt,
  model = "grok-2-image-1212",
  format = c("file", "base64", "raw")[1],
  api_key = Sys.getenv("XAI_API_KEY"),
  base_url = "https://api.x.ai/v1",
  filename = "image.png",
  user = NULL,
  organization = NULL
)
}
\description{
Generate image with Grok (which uses Flux).
Note: limitation of the prompt is about 1000 characters.
}
