% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-helpers.R
\name{ai.get_ollama_models}
\alias{ai.get_ollama_models}
\title{List available local Ollama models}
\usage{
ai.get_ollama_models(models = NULL, size = NULL)
}
\description{
List available local Ollama models
}
