% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cbind_sparse_matrix}
\alias{cbind_sparse_matrix}
\title{Column-bind sparse matrices}
\usage{
cbind_sparse_matrix(m1, m2)
}
\arguments{
\item{m1}{First sparse matrix.}

\item{m2}{Second sparse matrix.}
}
\value{
Combined sparse matrix.
}
\description{
Column-bind sparse matrices
}
\keyword{internal}
