% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseTraitMatrix}
\alias{collapseTraitMatrix}
\title{Collapse expanded trait matrix}
\usage{
collapseTraitMatrix(Y)
}
\arguments{
\item{Y}{Expanded trait matrix with "=" columns.}
}
\value{
Collapsed matrix with category columns.
}
\description{
Collapse expanded trait matrix
}
\keyword{internal}
