% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{computeConsensusGeneStats}
\alias{computeConsensusGeneStats}
\title{Compute gene statistics with consensus colors per layer}
\usage{
computeConsensusGeneStats(cons)
}
\arguments{
\item{cons}{Consensus WGCNA result object.}
}
\value{
List of gene statistics per layer.
}
\description{
Compute gene statistics with consensus colors per layer
}
