% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{computeConsensusModuleEnrichment}
\alias{computeConsensusModuleEnrichment}
\title{Compute consensus enrichment by calculating overlapping enriched terms.}
\usage{
computeConsensusModuleEnrichment(
  cons,
  GMT,
  annot,
  methods = c("fisher", "gsetcor", "xcor"),
  min.genes = 3,
  ntop = 400
)
}
\description{
Compute consensus enrichment by calculating overlapping enriched terms.
}
