% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{computeDistinctMatrix}
\alias{computeDistinctMatrix}
\title{Compute consensus matrix from list of matrices. The consensus
matrix checks for consistent sign and minimal threshold for each
matrix. Optionally filters on consistent p-value.}
\usage{
computeDistinctMatrix(matlist, ydim, psig = 0.05, min.diff = 0.3, consmax = 0)
}
\description{
Compute consensus matrix from list of matrices. The consensus
matrix checks for consistent sign and minimal threshold for each
matrix. Optionally filters on consistent p-value.
}
\keyword{internal}
