% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{computeModuleEnrichment}
\alias{computeModuleEnrichment}
\title{Compute enrichment of each WGCNA module using various
methods. Handles single-type and multi-omics WGCNA objects.}
\usage{
computeModuleEnrichment(
  wgcna,
  GMT,
  multi = FALSE,
  methods = c("fisher", "gsetcor", "xcor"),
  ntop = 200,
  xtop = 100,
  annot = NULL,
  xref = NULL,
  min.genes = 3,
  min.rho = 0.8,
  filter = NULL
)
}
\description{
Compute enrichment of each WGCNA module using various
methods. Handles single-type and multi-omics WGCNA objects.
}
