% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_sparse_matrix}
\alias{cor_sparse_matrix}
\title{Calculate sparse correlation matrix handling missing values}
\usage{
cor_sparse_matrix(G, mat)
}
\arguments{
\item{G}{Sparse matrix containing gene sets}

\item{mat}{Matrix of values}
}
\value{
Correlation matrix between G and mat
}
\description{
Calculate sparse correlation matrix handling missing values
}
\details{
If mat has no missing values, calculates corr. using corSparse.
Otherwise computes column-wise correlations only using non-missing values.
}
