% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{correct_dot_diagram}
\alias{correct_dot_diagram}
\title{Correct common DOT diagram issues}
\usage{
correct_dot_diagram(diagram)
}
\arguments{
\item{diagram}{DOT diagram string.}
}
\value{
Corrected DOT diagram string.
}
\description{
Correct common DOT diagram issues
}
\keyword{internal}
