% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cortest}
\alias{cortest}
\title{Pairwise correlation test with p-values}
\usage{
cortest(X, Y)
}
\arguments{
\item{X}{Numeric matrix of variables.}

\item{Y}{Numeric matrix of variables.}
}
\value{
List with rho, pvalue, and n.
}
\description{
Pairwise correlation test with p-values
}
\keyword{internal}
