% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_diagram}
\alias{create_diagram}
\title{Create module diagram in DOT format}
\usage{
create_diagram(
  wgcna_report,
  ai_model,
  graph = NULL,
  rankdir = "TB",
  correct = TRUE,
  double.check = TRUE
)
}
\arguments{
\item{wgcna_report}{Report text string.}

\item{ai_model}{LLM model name string.}

\item{graph}{Optional graph template.}

\item{rankdir}{DOT layout direction.}

\item{correct}{Apply DOT corrections.}

\item{double.check}{Validate DOT with DiagrammeR.}
}
\value{
DOT diagram string.
}
\description{
Create module diagram in DOT format
}
