% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_infographic}
\alias{create_infographic}
\title{Given the output from create_report() this function creates
an infographic by calling Gemini3. The genAI model is given the
report and asked to adhere to the included or external given
diagram (in DOT format).}
\usage{
create_infographic(
  report,
  diagram = NULL,
  prompt = NULL,
  model = "gemini-3-pro-image-preview",
  filename = "infographic.png",
  api_key = Sys.getenv("GEMINI_API_KEY")
)
}
\arguments{
\item{report}{Report text string.}

\item{diagram}{Optional DOT diagram string.}

\item{prompt}{Optional custom prompt.}

\item{model}{Gemini model name.}

\item{filename}{Output PNG filename.}
}
\value{
Path to output image file.
}
\description{
Given the output from create_report() this function creates
an infographic by calling Gemini3. The genAI model is given the
report and asked to adhere to the included or external given
diagram (in DOT format).
}
