% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{createConsensusLayers}
\alias{createConsensusLayers}
\title{Create aligned consensus WGCNA layers}
\usage{
createConsensusLayers(
  exprList,
  samples,
  contrasts = NULL,
  ngenes = 2000,
  power = 12,
  minModuleSize = 20,
  deepSplit = 2,
  mergeCutHeight = 0.15,
  minKME = 0.3,
  maxBlockSize = 9999,
  prefix = NULL,
  verbose = 1
)
}
\arguments{
\item{exprList}{List of expression matrices.}

\item{samples}{Sample metadata data frame.}

\item{contrasts}{Optional contrast matrix.}

\item{ngenes}{Maximum number of genes to use.}

\item{power}{Soft-thresholding power.}

\item{minModuleSize}{Minimum module size.}

\item{deepSplit}{Deep split sensitivity level.}

\item{mergeCutHeight}{Cut height for merging modules.}

\item{minKME}{Minimum KME to stay in module.}

\item{maxBlockSize}{Maximum block size for computation.}

\item{prefix}{Layer name prefixes.}

\item{verbose}{Verbosity level.}
}
\value{
List of aligned WGCNA layer results.
}
\description{
Create aligned consensus WGCNA layers
}
