% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{.computeModulesColorOrder}
\alias{.computeModulesColorOrder}
\title{Compute shared color order for module renaming}
\usage{
.computeModulesColorOrder(colors, unmergedColors)
}
\description{
Compute shared color order for module renaming
}
\keyword{internal}
