% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{.computeModulesCutTree}
\alias{.computeModulesCutTree}
\title{Determine module labels by cutting the gene tree}
\usage{
.computeModulesCutTree(
  geneTree,
  dissTOM,
  cutMethod,
  treeCut,
  treeCutCeiling,
  deepSplit,
  minModuleSize,
  minModuleSize2,
  verbose
)
}
\description{
Determine module labels by cutting the gene tree
}
\keyword{internal}
