% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{.computeModulesPower}
\alias{.computeModulesPower}
\title{Resolve soft-thresholding power for computeModules}
\usage{
.computeModulesPower(datExpr, power)
}
\description{
Resolve soft-thresholding power for computeModules
}
\keyword{internal}
