% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{.computeModulesTOM}
\alias{.computeModulesTOM}
\title{Compute TOM matrix using the requested strategy}
\usage{
.computeModulesTOM(
  datExpr,
  power,
  networkType,
  calcMethod,
  TOMType,
  lowrank,
  verbose
)
}
\description{
Compute TOM matrix using the requested strategy
}
\keyword{internal}
