% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getConsensusTopGenesAndSets}
\alias{.getConsensusTopGenesAndSets}
\title{Get consensus top genes and sets}
\usage{
.getConsensusTopGenesAndSets(
  cons,
  annot = NULL,
  module = NULL,
  ntop = 40,
  level = c("gene", "geneset")[1],
  rename = "symbol"
)
}
\arguments{
\item{cons}{Consensus WGCNA object.}

\item{annot}{Annotation table or NULL.}

\item{module}{Module names to select.}

\item{ntop}{Number of top entries.}

\item{level}{Feature level: "gene" or "geneset".}

\item{rename}{Column name for renaming.}
}
\value{
List with top sets, genes, and pheno.
}
\description{
Get consensus top genes and sets
}
\keyword{internal}
