% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{dot.rankdir}
\alias{dot.rankdir}
\title{Change layout of DOT string object}
\usage{
dot.rankdir(dot, dir)
}
\arguments{
\item{dot}{DOT format string.}

\item{dir}{Direction: "TB" or "LR".}
}
\value{
Modified DOT string.
}
\description{
Change layout of DOT string object
}
\keyword{internal}
