% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandPhenoMatrix}
\alias{expandPhenoMatrix}
\title{Expand phenotype matrix to binary design}
\usage{
expandPhenoMatrix(M, drop.ref = TRUE, keep.numeric = FALSE, check = TRUE)
}
\arguments{
\item{M}{Phenotype data.frame or matrix.}

\item{drop.ref}{Drop reference level columns.}

\item{keep.numeric}{Keep numeric columns as-is.}

\item{check}{Perform type checking.}
}
\value{
Expanded binary design matrix.
}
\description{
Expand phenotype matrix to binary design
}
\keyword{internal}
