% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{fastTOMsimilarity}
\alias{fastTOMsimilarity}
\title{Faster implementation of TOM computation using low-rank SVD approximation.}
\usage{
fastTOMsimilarity(A, tomtype = "signed", lowrank = 20)
}
\arguments{
\item{A}{Adjacency matrix.}

\item{tomtype}{TOM type (\"signed\" or \"unsigned\").}

\item{lowrank}{Low-rank approximation dimension.}
}
\value{
TOM similarity matrix.
}
\description{
Faster implementation of TOM computation using low-rank SVD approximation.
}
\keyword{internal}
