% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterColors}
\alias{filterColors}
\title{Filter module colors by KME}
\usage{
filterColors(
  X,
  colors,
  minKME = 0.3,
  mergeCutHeight = 0.15,
  minmodsize = 20,
  ntop = -1
)
}
\arguments{
\item{X}{Numeric expression matrix.}

\item{colors}{Module color assignments.}

\item{minKME}{Minimum KME threshold.}

\item{mergeCutHeight}{Cut height for merging modules.}

\item{minmodsize}{Minimum module size.}

\item{ntop}{Maximum features per module.}
}
\value{
Filtered color assignment vector.
}
\description{
Filter color vector by minimum KME and mergeCutHeight. Set color of
features with KME smaller than minKME to grey (or 0) group. Merge
similar modules with (module) correlation larger than
(1-mergeCutHeight) together.
}
