% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_modTraits}
\alias{get_modTraits}
\title{Get module-trait correlation data}
\usage{
get_modTraits(wgcna)
}
\arguments{
\item{wgcna}{A WGCNA result object.}
}
\value{
Numeric module-trait correlation matrix.
}
\description{
Get module-trait correlation data
}
\keyword{internal}
