% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getConsensusGeneStats}
\alias{getConsensusGeneStats}
\title{Get consensus gene statistics across layers}
\usage{
getConsensusGeneStats(cons, stats, trait, module = NULL)
}
\arguments{
\item{cons}{Consensus WGCNA result object.}

\item{stats}{Precomputed consensus stats list.}

\item{trait}{Trait name to extract.}

\item{module}{Module name to filter or NULL.}
}
\value{
List with consensus and full data frames.
}
\description{
Get consensus gene statistics across layers
}
