% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{getModuleCrossGenes}
\alias{getModuleCrossGenes}
\title{Get top correlated genes across modules}
\usage{
getModuleCrossGenes(
  wgcna,
  ref = NULL,
  ngenes = 100,
  multi = TRUE,
  modules = NULL
)
}
\arguments{
\item{wgcna}{WGCNA result object or list.}

\item{ref}{Reference layer name.}

\item{ngenes}{Number of top genes per module.}

\item{multi}{Use multi-layer mode.}

\item{modules}{Subset of modules to query.}
}
\value{
List of data frames with cross-module genes.
}
\description{
Get top correlated genes across modules
}
\keyword{internal}
