% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTopGenesAndSets}
\alias{getTopGenesAndSets}
\title{Get top genes and gene sets per module}
\usage{
getTopGenesAndSets(
  wgcna,
  annot = NULL,
  module = NULL,
  ntop = 40,
  psig = 0.05,
  level = "gene",
  rename = "symbol"
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{annot}{Annotation table or NULL.}

\item{module}{Module names to select.}

\item{ntop}{Number of top entries.}

\item{psig}{P-value significance threshold.}

\item{level}{Feature level: "gene" or "geneset".}

\item{rename}{Column name for renaming.}
}
\value{
List with top sets, genes, and pheno.
}
\description{
Get top genes and gene sets per module
}
\keyword{internal}
