% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{graph2dot}
\alias{graph2dot}
\title{Create DOT string object from graph object for sending to
LLM. Clean unneeded attributes.}
\usage{
graph2dot(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
DOT format string.
}
\description{
Create DOT string object from graph object for sending to
LLM. Clean unneeded attributes.
}
\keyword{internal}
