% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gset.fisherLIST}
\alias{gset.fisherLIST}
\title{Perform Fisher's exact test on gene sets, returning per-geneset results as a list}
\usage{
gset.fisherLIST(
  genes,
  genesets,
  background = NULL,
  fdr = 0.05,
  mc = TRUE,
  sort.by = "p.value",
  nmin = 3,
  min.genes = 15,
  max.genes = 500,
  method = "fast.fisher",
  check.background = TRUE,
  report.genes = FALSE,
  no.pass = NA,
  verbose = 1
)
}
\description{
Perform Fisher's exact test on gene sets, returning per-geneset results as a list
}
