% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAplus-package.R
\name{info}
\alias{info}
\title{Log info message with timestamp}
\usage{
info(..., type = "INFO")
}
\arguments{
\item{...}{Message parts to concatenate.}

\item{type}{Log level label string.}
}
\value{
NULL (called for side effect).
}
\description{
Log info message with timestamp
}
\keyword{internal}
