% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Date}
\alias{is.Date}
\title{Check if value is a Date}
\usage{
is.Date(x)
}
\arguments{
\item{x}{Value to test.}
}
\value{
Logical TRUE if Date parseable.
}
\description{
Check if value is a Date
}
\keyword{internal}
