% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logCPM}
\alias{logCPM}
\title{Log counts per million normalization}
\usage{
logCPM(counts, total = 1e+06, prior = 1, log = TRUE)
}
\arguments{
\item{counts}{Count matrix (dense or sparse).}

\item{total}{Library size normalization target.}

\item{prior}{Pseudocount before log transform.}

\item{log}{Whether to log2-transform.}
}
\value{
Normalized expression matrix.
}
\description{
Log counts per million normalization
}
\keyword{internal}
