% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeContrastsFromLabelMatrix}
\alias{makeContrastsFromLabelMatrix}
\title{Make contrast matrix from label matrix}
\usage{
makeContrastsFromLabelMatrix(lab.matrix)
}
\arguments{
\item{lab.matrix}{Label matrix with "_vs_" column names.}
}
\value{
Numeric contrast matrix.
}
\description{
Make contrast matrix from label matrix
}
\keyword{internal}
