% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat2gmt}
\alias{mat2gmt}
\title{Convert binary matrix to GMT list
This function converts binary matrix to a GMT (Gene Matrix
Transposed) list, The binary matrix indicates presence or absence of
genes in each gene set: rows are genes and columns are gene sets.}
\usage{
mat2gmt(mat)
}
\arguments{
\item{mat}{Matrix with non-zero entries representing genes in
each gene set where rows are genes and columns are  gene sets.}
}
\value{
List of vector representing each gene set. Each list
  element correspond to a gene set and is a vector of genes
}
\description{
Convert binary matrix to GMT list
This function converts binary matrix to a GMT (Gene Matrix
Transposed) list, The binary matrix indicates presence or absence of
genes in each gene set: rows are genes and columns are gene sets.
}
