% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{matchColors}
\alias{matchColors}
\title{Match module colors to reference labels}
\usage{
matchColors(wgcna, refcolors)
}
\arguments{
\item{wgcna}{WGCNA result object.}

\item{refcolors}{Reference color assignments.}
}
\value{
Updated WGCNA object with matched colors.
}
\description{
Match module colors to reference labels
}
\keyword{internal}
