% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAplus-package.R
\name{mem.proc}
\alias{mem.proc}
\title{Get process virtual memory usage}
\usage{
mem.proc(digits = 0)
}
\arguments{
\item{digits}{Number of decimal places.}
}
\value{
Memory usage string in MB.
}
\description{
Get process virtual memory usage
}
\keyword{internal}
