% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAplus-package.R
\name{mem.vmrss}
\alias{mem.vmrss}
\title{Get VM RSS memory usage}
\usage{
mem.vmrss(digits = 0)
}
\arguments{
\item{digits}{Number of decimal places.}
}
\value{
Memory usage string in MB.
}
\description{
Get VM RSS memory usage
}
\keyword{internal}
