% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{merge_block_dendrograms}
\alias{merge_block_dendrograms}
\title{Merge block dendrograms into single tree}
\usage{
merge_block_dendrograms(net, X, method = 1)
}
\arguments{
\item{net}{Network object with dendrograms.}

\item{X}{Data matrix for block means.}

\item{method}{Merging method (1 or 2).}
}
\value{
Merged hclust dendrogram object.
}
\description{
Merge block dendrograms into single tree
}
\keyword{internal}
