% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_sparse_matrix}
\alias{merge_sparse_matrix}
\title{Merge two sparse matrices}
\usage{
merge_sparse_matrix(m1, m2, margin = NULL, verbose = 1)
}
\arguments{
\item{m1}{First sparse matrix (or NULL).}

\item{m2}{Second sparse matrix (or NULL).}

\item{margin}{Unused, reserved for future.}

\item{verbose}{Verbosity level.}
}
\value{
Combined sparse matrix.
}
\description{
Merge two sparse matrices
}
\keyword{internal}
