% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{mergeCloseModules}
\alias{mergeCloseModules}
\title{Merge modules with similar eigengenes}
\usage{
mergeCloseModules(datExpr, colors, cutHeight, MEs = NULL)
}
\arguments{
\item{datExpr}{Expression data (samples x genes).}

\item{colors}{Module color assignments.}

\item{cutHeight}{Cut height for merging.}

\item{MEs}{Pre-computed module eigengenes.}
}
\value{
List with merged colors and eigengenes.
}
\description{
Merge modules with similar eigengenes
}
\keyword{internal}
