% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeME}
\alias{mergeME}
\title{Merge multiple ME matrices into one. Allow different dimensions.}
\usage{
mergeME(mlist, me2 = NULL, prefix = FALSE)
}
\arguments{
\item{mlist}{List of ME matrices.}

\item{me2}{Optional second ME matrix.}

\item{prefix}{Prefix columns with list names.}
}
\value{
Merged eigengene matrix.
}
\description{
Merge multiple ME matrices into one. Allow different dimensions.
}
\keyword{internal}
