% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mofa.get_prefix}
\alias{mofa.get_prefix}
\title{Extract MOFA prefix from names}
\usage{
mofa.get_prefix(x)
}
\arguments{
\item{x}{Character vector, matrix, or data.frame.}
}
\value{
Character vector of extracted prefixes.
}
\description{
Extract MOFA prefix from names
}
\keyword{internal}
