% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mofa.merge_data}
\alias{mofa.merge_data}
\title{Merge MOFA data layers by row}
\usage{
mofa.merge_data(xx)
}
\arguments{
\item{xx}{Named list of data matrices.}
}
\value{
Combined matrix with prefixed rownames.
}
\description{
Merge MOFA data layers by row
}
\keyword{internal}
