% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mofa.merge_data2}
\alias{mofa.merge_data2}
\title{Merge MOFA data with flexible options}
\usage{
mofa.merge_data2(xdata, merge.rows = "prefix", merge.cols = "union")
}
\arguments{
\item{xdata}{Named list of data matrices.}

\item{merge.rows}{Row merge strategy: "prefix", "union", or "intersect".}

\item{merge.cols}{Column merge strategy: "prefix", "union", or "intersect".}
}
\value{
Merged numeric matrix.
}
\description{
Merge MOFA data with flexible options
}
\keyword{internal}
