% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mofa.prefix}
\alias{mofa.prefix}
\title{Add MOFA feature prefix}
\usage{
mofa.prefix(xx)
}
\arguments{
\item{xx}{Named list of matrices or vectors.}
}
\value{
Input with layer-name prefixes added.
}
\description{
Add MOFA feature prefix
}
\keyword{internal}
