% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mofa.split_data}
\alias{mofa.split_data}
\title{Split MOFA data by prefix}
\usage{
mofa.split_data(X, keep.prefix = FALSE)
}
\arguments{
\item{X}{Matrix with prefixed rownames.}

\item{keep.prefix}{Whether to keep prefix in names.}
}
\value{
Named list of matrices per layer.
}
\description{
Split MOFA data by prefix
}
\keyword{internal}
