% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mofa.strip_prefix}
\alias{mofa.strip_prefix}
\title{Strip MOFA feature prefix}
\usage{
mofa.strip_prefix(xx)
}
\arguments{
\item{xx}{Character vector, matrix, or list.}
}
\value{
Input with prefixes removed.
}
\description{
Strip MOFA feature prefix
}
\keyword{internal}
