% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mofa.topSD}
\alias{mofa.topSD}
\title{Select top SD features per MOFA layer}
\usage{
mofa.topSD(xdata, ntop)
}
\arguments{
\item{xdata}{Matrix or list of matrices.}

\item{ntop}{Number of top features to keep.}
}
\value{
Filtered data with top-variance features.
}
\description{
Select top SD features per MOFA layer
}
\keyword{internal}
