% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pgx.supercell}
\alias{pgx.supercell}
\title{Compute supercells for single-cell data}
\usage{
pgx.supercell(
  counts,
  meta,
  group = NULL,
  gamma = 20,
  nvargenes = 1000,
  log.transform = TRUE
)
}
\arguments{
\item{counts}{Count matrix (genes x cells).}

\item{meta}{Cell metadata data.frame.}

\item{group}{Grouping variable name or vector.}

\item{gamma}{Graining level for SuperCell.}

\item{nvargenes}{Number of variable genes to use.}

\item{log.transform}{Whether to log-CPM transform.}
}
\value{
List with counts, meta, and membership.
}
\description{
Compute supercells for single-cell data
}
\keyword{internal}
