% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pickSoftThreshold}
\alias{pickSoftThreshold}
\title{Pick soft thresholding power
Better (?) method to pick soft threshold (aka power).}
\usage{
pickSoftThreshold(
  datExpr,
  sft = NULL,
  rcut = 0.85,
  method = c("sft", "iqr")[1],
  nmax = -1,
  powers = NULL,
  verbose = 1
)
}
\arguments{
\item{datExpr}{Numeric expression data matrix.}

\item{sft}{Pre-computed soft threshold result.}

\item{rcut}{R-squared cutoff for model fit.}

\item{method}{Selection method: "sft" or "iqr".}

\item{nmax}{Maximum features to subsample.}

\item{powers}{Numeric vector of powers to test.}

\item{verbose}{Verbosity level.}
}
\value{
Integer optimal soft-thresholding power.
}
\description{
Pick soft thresholding power
Better (?) method to pick soft threshold (aka power).
}
