% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-multi.R
\name{plotConsensusOverlapHeatmap}
\alias{plotConsensusOverlapHeatmap}
\title{Plot consensus module overlap heatmap}
\usage{
plotConsensusOverlapHeatmap(
  net1,
  net2,
  setLabels = NULL,
  lab.line = c(8, 8),
  plotDendro = FALSE,
  setpar = TRUE
)
}
\arguments{
\item{net1}{First network object.}

\item{net2}{Second network object.}

\item{setLabels}{Labels for the two sets.}

\item{lab.line}{Margin line for axis labels.}

\item{plotDendro}{If TRUE, plot dendrograms above.}

\item{setpar}{If TRUE, set par layout.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot consensus module overlap heatmap
}
