% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-eigengene.R
\name{plotConsensusSampleDendroAndColors}
\alias{plotConsensusSampleDendroAndColors}
\title{Plot consensus sample dendrogram with colors}
\usage{
plotConsensusSampleDendroAndColors(
  cons,
  i,
  what = c("both", "me", "traits")[1],
  show.me = TRUE,
  show.traits = TRUE,
  show.contrasts = TRUE,
  clust.expr = TRUE,
  setLayout = TRUE,
  marAll = c(0.2, 7, 1.5, 0.5),
  colorHeightMax = 0.6,
  main = NULL
)
}
\arguments{
\item{cons}{A consensus WGCNA result object.}

\item{i}{Index of consensus layer.}

\item{what}{What to show: "both", "me", or "traits".}

\item{show.me}{Show module eigengenes.}

\item{show.traits}{Show trait values.}

\item{show.contrasts}{Show contrast values.}

\item{clust.expr}{Cluster by expression similarity.}

\item{setLayout}{Whether to set layout.}

\item{marAll}{Margin sizes vector.}

\item{colorHeightMax}{Maximum color row height.}

\item{main}{Plot title string.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot consensus sample dendrogram with colors
}
