% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotConsensusTraitCorrelation}
\alias{plotConsensusTraitCorrelation}
\title{Plot trait correlation scatter}
\usage{
plotConsensusTraitCorrelation(cons, traits = NULL)
}
\description{
Plot trait correlation scatter
}
