% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dendro.R
\name{plotDendroAndColors_multi}
\alias{plotDendroAndColors_multi}
\title{Plot dendrogram with traits for multiple datasets}
\usage{
plotDendroAndColors_multi(
  multi,
  show.traits = TRUE,
  show.contrasts = TRUE,
  traits = NULL,
  main = NULL,
  rm.na = TRUE,
  use.tree = 1,
  marAll = c(0.2, 8, 2, 0.2),
  setLayout = TRUE,
  ...
)
}
\arguments{
\item{multi}{Named list of WGCNA objects.}

\item{show.traits}{Show trait correlations.}

\item{show.contrasts}{Show contrast correlations.}

\item{traits}{Traits to include.}

\item{main}{Plot title string.}

\item{rm.na}{Remove NA-only columns.}

\item{use.tree}{Which tree layer to use.}

\item{marAll}{Margin sizes vector.}

\item{setLayout}{Whether to set layout.}

\item{...}{Additional arguments passed on.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot dendrogram with traits for multiple datasets
}
