% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dendro.R
\name{plotDendroAndTraitCorrelation}
\alias{plotDendroAndTraitCorrelation}
\title{Plot dendrogram with trait correlations (deprecated)}
\usage{
plotDendroAndTraitCorrelation(
  wgcna,
  traits = NULL,
  show.traits = TRUE,
  show.contrasts = TRUE,
  show.kme = FALSE,
  main = NULL,
  block = NULL,
  rm.na = TRUE,
  use.tree = 0,
  marAll = c(0.2, 8, 2, 0.2),
  setLayout = TRUE,
  ...
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{traits}{Traits to include.}

\item{show.traits}{Show trait correlations.}

\item{show.contrasts}{Show contrast correlations.}

\item{show.kme}{Show eigengene correlations.}

\item{main}{Plot title string.}

\item{block}{Block number to plot.}

\item{rm.na}{Remove NA-only columns.}

\item{use.tree}{Tree index for consensus.}

\item{marAll}{Margin sizes vector.}

\item{setLayout}{Whether to set layout.}

\item{...}{Additional arguments passed on.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot dendrogram with trait correlations (deprecated)
}
