% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-eigengene.R
\name{plotEigenGeneClusterDendrogram}
\alias{plotEigenGeneClusterDendrogram}
\title{Plot eigengene cluster dendrogram}
\usage{
plotEigenGeneClusterDendrogram(
  wgcna = NULL,
  ME = NULL,
  add_traits = TRUE,
  horiz = FALSE,
  setMargins = TRUE,
  method = "wgcna",
  showlabels = TRUE,
  plot = TRUE,
  multi = FALSE,
  main = NULL
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{ME}{Eigengene matrix to use directly.}

\item{add_traits}{Include traits in clustering.}

\item{horiz}{Plot dendrogram horizontally.}

\item{setMargins}{Adjust plot margins automatically.}

\item{method}{Clustering method to use.}

\item{showlabels}{Show dendrogram labels.}

\item{plot}{Whether to generate the plot.}

\item{multi}{Use multi-dataset mode.}

\item{main}{Plot title string.}
}
\value{
Hierarchical clustering object (invisible).
}
\description{
Plot eigengene cluster dendrogram
}
