% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-eigengene.R
\name{plotEigenGeneGraph}
\alias{plotEigenGeneGraph}
\title{Plot eigengene network as graph}
\usage{
plotEigenGeneGraph(
  wgcna,
  add_traits = TRUE,
  main = NULL,
  multi = FALSE,
  vcex = 1,
  labcex = 1,
  mincor = 0.5,
  as.phylo = TRUE
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{add_traits}{Include traits in graph.}

\item{main}{Plot title string.}

\item{multi}{Use multi-dataset mode.}

\item{vcex}{Vertex size scaling factor.}

\item{labcex}{Label size scaling factor.}

\item{mincor}{Minimum correlation for non-phylo edge (ignored when as.phylo=TRUE).}

\item{as.phylo}{If TRUE, use phylo-based graph layout (Kamada-Kawai).}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot eigengene network as graph
}
