% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotFeatureUMAP}
\alias{plotFeatureUMAP}
\title{Plot feature UMAP colored by module}
\usage{
plotFeatureUMAP(
  wgcna,
  nhub = 3,
  method = "clust",
  scale = FALSE,
  main = NULL,
  plotlib = "base",
  annot = NULL
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{nhub}{Number of hub genes to label.}

\item{method}{Embedding method: "clust", "umap", or "mds".}

\item{scale}{If TRUE, scale features.}

\item{main}{Plot title.}

\item{plotlib}{Plotting library to use.}

\item{annot}{Annotation table for gene symbols.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot feature UMAP colored by module
}
