% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotGeneNetwork}
\alias{plotGeneNetwork}
\title{Plot gene correlation network graph}
\usage{
plotGeneNetwork(
  wgcna,
  genes,
  col = NULL,
  edge.alpha = 0.3,
  rgamma = 4,
  edge.width = 6,
  alpha = 0.5,
  min.rho = 0.5,
  setpar = TRUE
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{genes}{Gene names to include.}

\item{col}{Vertex colors.}

\item{edge.alpha}{Edge transparency.}

\item{rgamma}{Gamma exponent for edge weights.}

\item{edge.width}{Base edge width.}

\item{alpha}{Vertex color transparency.}

\item{min.rho}{Minimum correlation threshold.}

\item{setpar}{If TRUE, set par layout.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot gene correlation network graph
}
