% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot MDS of features colored by module}
\usage{
plotMDS(wgcna, main = NULL, scale = FALSE)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{main}{Plot title.}

\item{scale}{If TRUE, scale features.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot MDS of features colored by module
}
