% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{plotModuleHeatmap}
\alias{plotModuleHeatmap}
\title{Plot module expression or correlation heatmap}
\usage{
plotModuleHeatmap(
  wgcna,
  module,
  genes = NULL,
  rgamma = 4,
  min.rho = 0,
  cex = 0.8,
  nmax = -1,
  cluster = TRUE,
  type = c("expression", "correlation")[1],
  heatmap.mar = c(7, 7),
  main = NULL
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{module}{Module name to plot.}

\item{genes}{Gene names to include.}

\item{rgamma}{Gamma exponent for correlation.}

\item{min.rho}{Minimum correlation threshold.}

\item{cex}{Text size for labels.}

\item{nmax}{Max number of genes to show.}

\item{cluster}{If TRUE, cluster genes.}

\item{type}{Type: "expression" or "correlation".}

\item{heatmap.mar}{Heatmap margins as c(bottom, left).}

\item{main}{Plot title.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot module expression or correlation heatmap
}
