% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotModuleHubGenes}
\alias{plotModuleHubGenes}
\title{Plot hub gene network per module}
\usage{
plotModuleHubGenes(wgcna, modules = NULL, alpha = 0.5, setpar = TRUE)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{modules}{Module names to plot.}

\item{alpha}{Color transparency for vertices.}

\item{setpar}{If TRUE, set par layout.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot hub gene network per module
}
