% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotModuleScores.BAK}
\alias{plotModuleScores.BAK}
\title{Plot top modules most correlated with trait for multi expression
data.}
\usage{
plotModuleScores.BAK(
  res,
  trait,
  multi = FALSE,
  nmax = 16,
  collapse.trait = FALSE,
  base_size = 18,
  plotlib = "base",
  setpar = TRUE
)
}
\arguments{
\item{res}{A WGCNA or multi-WGCNA result object.}

\item{trait}{Trait name to plot.}

\item{multi}{If TRUE, treat \code{res} as a multi-dataset object.}

\item{nmax}{Maximum number of modules to plot.}

\item{collapse.trait}{If TRUE, collapse trait matrix (multi only).}

\item{setpar}{Set par layout: TRUE/1 for portrait, 2 for landscape, FALSE for none.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Works with both single-dataset and multi-dataset objects.
When \code{multi = TRUE}, the object is expected to have a list of
ME matrices in \code{res$MEs} and trait data are merged across datasets.
}
