% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotModuleSignificance}
\alias{plotModuleSignificance}
\title{Plot module significance.}
\usage{
plotModuleSignificance(wgcna, trait, main = NULL, abs = FALSE)
}
\description{
Plot module significance.
}
