% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{plotModuleTraitHeatmap}
\alias{plotModuleTraitHeatmap}
\title{Plot module-trait correlation heatmap}
\usage{
plotModuleTraitHeatmap(
  wgcna,
  setpar = TRUE,
  cluster = FALSE,
  multi = FALSE,
  main = NULL,
  justdata = FALSE,
  transpose = FALSE,
  colorlabel = TRUE,
  show = c("both", "traits", "contrasts")[1],
  nmax = -1,
  tmax = -1,
  text = TRUE,
  pstar = TRUE
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{setpar}{If TRUE, set par margins.}

\item{cluster}{If TRUE, cluster rows and columns.}

\item{multi}{If TRUE, treat as multi-dataset.}

\item{main}{Plot title.}

\item{justdata}{If TRUE, return data only.}

\item{transpose}{If TRUE, transpose the matrix.}

\item{colorlabel}{If TRUE, use color labels.}

\item{show}{Show "both", "traits", or "contrasts".}

\item{nmax}{Max number of modules to show.}

\item{tmax}{Max number of traits to show.}

\item{text}{If TRUE, display correlation values.}

\item{pstar}{If TRUE, show significance stars.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot module-trait correlation heatmap
}
