% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-eigengene.R
\name{plotMultiEigengeneCorrelation}
\alias{plotMultiEigengeneCorrelation}
\title{Plot inter-dataset eigengene correlation heatmaps}
\usage{
plotMultiEigengeneCorrelation(
  wgcna,
  addtraits = TRUE,
  phenotype = NULL,
  nmax = -1,
  main = NULL,
  showvalues = FALSE,
  showsig = TRUE,
  cex.text = 0.7,
  cex.lab = 0.8,
  fixcluster = TRUE,
  setpar = TRUE
)
}
\arguments{
\item{wgcna}{Named list of WGCNA objects.}

\item{addtraits}{Include traits in correlation.}

\item{phenotype}{Phenotype for conditioning.}

\item{nmax}{Maximum number of eigengenes shown.}

\item{main}{Plot title string.}

\item{showvalues}{Show correlation values.}

\item{showsig}{Show significance stars.}

\item{cex.text}{Text character expansion factor.}

\item{cex.lab}{Label character expansion factor.}

\item{fixcluster}{Fix cluster ordering.}

\item{setpar}{Set plotting parameters.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot inter-dataset eigengene correlation heatmaps
}
