% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-multi.R
\name{plotPreservationModuleTraits}
\alias{plotPreservationModuleTraits}
\title{Plot preservation and module-trait heatmaps}
\usage{
plotPreservationModuleTraits(
  pres,
  subplots = c("zsummary", "consmt", "wt.consmt"),
  order.by = "name",
  setpar = TRUE,
  rm.na = FALSE
)
}
\arguments{
\item{pres}{Preservation results object.}

\item{subplots}{Subplots to show: "zsummary", "consmt", "wt.consmt".}

\item{order.by}{Order modules by "name", "zsummary", or "clust".}

\item{setpar}{If TRUE, set par layout.}

\item{rm.na}{If TRUE, remove all-NA columns.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot preservation and module-trait heatmaps
}
