% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-multi.R
\name{plotPreservationSummaries}
\alias{plotPreservationSummaries}
\title{Plot module preservation summary statistics}
\usage{
plotPreservationSummaries(pres, setpar = TRUE)
}
\arguments{
\item{pres}{Preservation results object.}

\item{setpar}{If TRUE, set par layout.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot module preservation summary statistics
}
