% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{plotTOM}
\alias{plotTOM}
\title{Plot TOM heatmap with module colors}
\usage{
plotTOM(
  wgcna,
  justdata = FALSE,
  block = NULL,
  legend = TRUE,
  downsample = NULL
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{justdata}{If TRUE, return dissimilarity matrix only.}

\item{block}{Block number to plot.}

\item{legend}{If TRUE, add color legend.}

\item{downsample}{Downsample to this many genes.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot TOM heatmap with module colors
}
