% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotTopModules}
\alias{plotTopModules}
\title{Plot top modules most correlated with trait.}
\usage{
plotTopModules(wgcna, trait, nmax = 16, setpar = TRUE)
}
\description{
Plot top modules most correlated with trait.
}
