% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotTopModules_multi}
\alias{plotTopModules_multi}
\title{Plot top modules most correlated with trait for multi expression
data.}
\usage{
plotTopModules_multi(
  multi,
  trait,
  nmax = 16,
  collapse = FALSE,
  plotlib = "base",
  setpar = TRUE
)
}
\description{
Plot top modules most correlated with trait for multi expression
data.
}
