% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotTraitCorrelationBarPlots}
\alias{plotTraitCorrelationBarPlots}
\title{Plot trait correlation as bar plots}
\usage{
plotTraitCorrelationBarPlots(
  res,
  trait,
  multi = FALSE,
  colored = TRUE,
  beside = TRUE,
  main = NULL,
  cex.main = 1.3,
  setpar = TRUE
)
}
\arguments{
\item{res}{A WGCNA result object.}

\item{trait}{Trait name(s) to plot.}

\item{multi}{If TRUE, use multi-dataset results.}

\item{colored}{If TRUE, color bars by module.}

\item{beside}{If TRUE, draw bars side-by-side.}

\item{main}{Plot title.}

\item{cex.main}{Title text size.}

\item{setpar}{If TRUE, set par layout.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot trait correlation as bar plots
}
