% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probe2symbol}
\alias{probe2symbol}
\title{Convert probe IDs to gene symbols}
\usage{
probe2symbol(
  probes,
  annot_table,
  query = "symbol",
  key = NULL,
  fill_na = FALSE,
  add_datatype = FALSE
)
}
\arguments{
\item{probes}{Character vector of probe IDs.}

\item{annot_table}{Annotation data.frame with mappings.}

\item{query}{Target column name to return.}

\item{key}{Column name to match probes.}

\item{fill_na}{Fill missing with probe IDs.}

\item{add_datatype}{Prepend data type prefix.}
}
\value{
Character vector of mapped symbols.
}
\description{
Convert probe IDs to gene symbols
}
\keyword{internal}
