% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{purpleGreyYellow}
\alias{purpleGreyYellow}
\title{Purple-grey-yellow color palette}
\usage{
purpleGreyYellow(n)
}
\arguments{
\item{n}{Number of colors to generate.}
}
\value{
Character vector of hex colors.
}
\description{
Purple-grey-yellow color palette
}
\keyword{internal}
