% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_enrichment_methods}
\alias{run_enrichment_methods}
\title{Run enrichment analysis methods on modules}
\usage{
run_enrichment_methods(
  ME,
  me.genes,
  GMT,
  geneX,
  methods = c("fisher", "gsetcor", "xcor"),
  ntop = 400,
  xtop = 100,
  min.genes = 3,
  min.rho = 0.8
)
}
\arguments{
\item{ME}{Module eigengene matrix.}

\item{me.genes}{List of gene vectors per module.}

\item{GMT}{Gene-set membership sparse matrix.}

\item{geneX}{Gene expression matrix (genes x samples).}

\item{methods}{Character vector of methods to run.}

\item{ntop}{Maximum gene sets to return.}

\item{xtop}{Maximum cross-correlated features to add.}

\item{min.genes}{Minimum genes per gene set.}

\item{min.rho}{Minimum correlation for cross-features.}
}
\value{
List of enrichment data frames per module.
}
\description{
Run enrichment analysis methods on modules
}
\keyword{internal}
