% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaleTOMs}
\alias{scaleTOMs}
\title{Scale TOM matrices to equal quantiles}
\usage{
scaleTOMs(TOMs, scaleP = 0.95)
}
\arguments{
\item{TOMs}{List of TOM matrices.}

\item{scaleP}{Reference quantile for scaling.}
}
\value{
List of scaled TOM matrices.
}
\description{
Scale a list of TOM matrices so that the quantiles (default p=0.95)
are equal after scaling with respect to the first TOM matrix.
}
\keyword{internal}
