% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy.dataframe}
\alias{tidy.dataframe}
\title{Tidy a dataframe with type inference}
\usage{
tidy.dataframe(Y)
}
\arguments{
\item{Y}{Data.frame or matrix to tidy.}
}
\value{
Data.frame with inferred column types.
}
\description{
Tidy a dataframe with type inference
}
\keyword{internal}
