% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tomclust}
\alias{tomclust}
\title{TOM-based hierarchical clustering}
\usage{
tomclust(X, power = 6)
}
\arguments{
\item{X}{Numeric expression matrix.}

\item{power}{Soft-thresholding power.}
}
\value{
An hclust dendrogram object.
}
\description{
Wrapper to hclust from matrix using default WGCNA parameters.
}
\keyword{internal}
