% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lasagna.R
\name{wgcna.plot_lasagna}
\alias{wgcna.plot_lasagna}
\title{Plot Lasagna graph with nodes colered by WGCNA ME colors}
\usage{
wgcna.plot_lasagna(
  wgcna,
  pheno,
  layout = "tsne",
  edge_color = c("blue", "magenta")
)
}
\description{
Plot Lasagna graph with nodes colered by WGCNA ME colors
}
