% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ai.ask_ellmer}
\alias{ai.ask_ellmer}
\title{Ask an LLM a question using the ellmer package}
\usage{
ai.ask_ellmer(question, model = DEFAULT_LLM, prompt = NULL)
}
\description{
Ask an LLM a question using the ellmer package
}
