% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-helpers.R
\name{ai.ask_tidyprompt}
\alias{ai.ask_tidyprompt}
\title{Ask an LLM a question using the tidyprompt package}
\usage{
ai.ask_tidyprompt(question, model, verbose = 0)
}
\description{
Ask an LLM a question using the tidyprompt package
}
