% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-helpers.R
\name{ai.create_image_gemini}
\alias{ai.create_image_gemini}
\title{Generate image with Gemini (aka Nano Banana).
Note this model handles very large prompts correctly.}
\usage{
ai.create_image_gemini(
  prompt,
  model = "gemini-2.5-flash-image",
  api_key = Sys.getenv("GEMINI_API_KEY"),
  format = c("file", "base64", "raw")[1],
  filename = "image.png",
  aspectRatio = "16:9",
  imageSize = "1K",
  base_url = "https://generativelanguage.googleapis.com/v1beta"
)
}
\description{
Generate image with Gemini (aka Nano Banana).
Note this model handles very large prompts correctly.
}
