% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{computeGeneStats}
\alias{computeGeneStats}
\title{Compute general feature statistics after WGCNA results.}
\usage{
computeGeneStats(net, datExpr = NULL, datTraits = NULL, TOM = NULL)
}
\arguments{
\item{net}{WGCNA network object with MEs and colors, or a full wgcna result object.}

\item{datExpr}{Expression data matrix (optional if net is a full wgcna object).}

\item{datTraits}{Trait data matrix (optional if net is a full wgcna object).}

\item{TOM}{Topological overlap matrix or NULL (optional if net is a full wgcna object).}
}
\value{
List of gene statistic matrices.
}
\description{
Compute general feature statistics after WGCNA results.
}
