% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{computeGeneStats}
\alias{computeGeneStats}
\title{Compute general feature statistics after WGCNA results.}
\usage{
computeGeneStats(net, datExpr, datTraits, TOM)
}
\arguments{
\item{net}{WGCNA network object with MEs and colors.}

\item{datExpr}{Expression data matrix.}

\item{datTraits}{Trait data matrix.}

\item{TOM}{Topological overlap matrix or NULL.}
}
\value{
List of gene statistic matrices.
}
\description{
Compute general feature statistics after WGCNA results.
}
