% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{getTopModules}
\alias{getTopModules}
\title{Get top correlated modules}
\usage{
getTopModules(wgcna, topratio = 0.85, kx = 4, rm.grey = TRUE, multi = FALSE)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{topratio}{Ratio threshold for top selection.}

\item{kx}{Power exponent for ranking.}

\item{rm.grey}{Remove grey module.}

\item{multi}{Use multi-omics mode.}
}
\value{
Character vector of top module names.
}
\description{
Get top correlated modules
}
