% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{getTopTables}
\alias{getTopTables}
\title{Get multi-dataset top genes and sets tables}
\usage{
getTopTables(
  wgcna,
  annot = NULL,
  module = NULL,
  psig = 0.05,
  ntop = 40,
  level = NULL,
  rename = "symbol"
)
}
\arguments{
\item{annot}{Annotation table or NULL.}

\item{module}{Module names to select.}

\item{psig}{P-value significance threshold.}

\item{ntop}{Number of top entries.}

\item{level}{Feature level or NULL.}

\item{rename}{Column name for renaming.}

\item{multi_wgcna}{Multi-omics WGCNA object.}
}
\value{
List with top sets, genes, and pheno.
}
\description{
Get multi-dataset top genes and sets tables
}
\keyword{internal}
