% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gset.rankcor}
\alias{gset.rankcor}
\title{Calculate gene set rank correlation
Compute rank correlation between a gene rank vector/matrix and gene sets}
\usage{
gset.rankcor(rnk, gset, compute.p = FALSE, use.rank = TRUE)
}
\arguments{
\item{rnk}{Numeric vector or matrix of gene ranks, with genes as row names}

\item{gset}{Numeric matrix of gene sets, with genes as row/column names}

\item{compute.p}{Logical indicating whether to compute p-values}

\item{use.rank}{Logical indicating whether to rank transform rnk before correlation}
}
\value{
Named list with components:
\itemize{
 \item rho - Matrix of correlation coefficients between rnk and gset
 \item p.value - Matrix of p-values for correlation (if compute.p = TRUE)
 \item q.value - Matrix of FDR adjusted p-values (if compute.p = TRUE)
}
}
\description{
Calculate gene set rank correlation
Compute rank correlation between a gene rank vector/matrix and gene sets
}
\details{
This function calculates sparse rank correlation between rnk and each
column of gset using \code{qlcMatrix::corSparse()}. It handles missing values in
rnk by computing column-wise correlations.
P-values are computed from statistical distribution
}
\examples{
\dontrun{
library(playbase)
ranks <- sample(1:10000, 1000, replace = TRUE)
names(ranks) <- replicate(1000, paste(sample(LETTERS, 4, replace = TRUE), collapse = ""))
genesets <- matrix(rnorm(1000 * 20), ncol = 20)
rownames(genesets) <- names(ranks)
gset.rankcor(ranks, genesets, compute.p = TRUE)
}
}
