% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labels2colors}
\alias{labels2colors}
\title{Convert numeric labels to colors}
\usage{
labels2colors(colors, ...)
}
\arguments{
\item{colors}{Numeric or character label vector.}

\item{...}{Additional arguments passed to WGCNA::labels2colors.}
}
\value{
Character vector of color names.
}
\description{
Converts WGCNA labels (numeric or color) to colors.
}
