% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dendro.R
\name{plotDendroAndColors}
\alias{plotDendroAndColors}
\title{Plot gene dendrogram with module colors}
\usage{
plotDendroAndColors(
  wgcna,
  main = NULL,
  extra.colors = NULL,
  show.kme = FALSE,
  show.traits = FALSE,
  show.contrasts = FALSE,
  show.mat = NULL,
  clust = TRUE,
  use.tree = 0,
  block = 1,
  rm.na = TRUE,
  sd.wt = 0,
  nmax = -1,
  marAll = c(0.4, 5, 1, 0.2),
  setLayout = TRUE,
  ...
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{main}{Plot title string.}

\item{extra.colors}{Additional color rows to display.}

\item{show.kme}{Show module eigengene correlations.}

\item{show.traits}{Show trait correlations.}

\item{show.contrasts}{Show contrast correlations.}

\item{show.mat}{Custom matrix to show.}

\item{clust}{Cluster columns by correlation.}

\item{use.tree}{Tree index for consensus.}

\item{block}{Block number to plot.}

\item{rm.na}{Remove NA-only columns.}

\item{sd.wt}{Standard deviation weighting factor.}

\item{nmax}{Maximum number of columns shown.}

\item{marAll}{Margin sizes vector.}

\item{setLayout}{Whether to set layout.}

\item{...}{Additional arguments to WGCNA::plotDendroAndColors.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot gene dendrogram with module colors
}
