% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-eigengene.R
\name{plotEigenGeneAdjacencyHeatmap}
\alias{plotEigenGeneAdjacencyHeatmap}
\title{Plot the adjacency correlation heatmap matrix of eigengenes with or
without traits. This can show how traits cluster together with the
eigengenes.}
\usage{
plotEigenGeneAdjacencyHeatmap(
  wgcna,
  add_traits = TRUE,
  traits = NULL,
  add_me = TRUE,
  marx = 1,
  main = NULL,
  multi = FALSE,
  phenotype = NULL,
  colorlabel = TRUE,
  text = FALSE,
  pstar = TRUE,
  power = 1,
  setMargins = TRUE,
  mar1 = c(5.6, 4.5, 1.8, 0),
  mar2 = c(8, 10, 4, 2),
  cex.lab = 0.8,
  cex.text = 0.7,
  plotDendro = TRUE,
  plotHeatmap = TRUE,
  dendro.horiz = TRUE,
  dendro.width = 0.3,
  dendro.labels = TRUE,
  nmax = -1,
  fixclust = FALSE,
  mask.intra = FALSE,
  justdata = FALSE
)
}
\arguments{
\item{wgcna}{A WGCNA result object.}

\item{add_traits}{Include traits in heatmap.}

\item{traits}{Specific traits to include.}

\item{add_me}{Include module eigengenes.}

\item{marx}{Margin expansion factor.}

\item{main}{Plot title string.}

\item{multi}{Use multi-dataset mode.}

\item{phenotype}{Phenotype for conditioning.}

\item{colorlabel}{Color-code row/column labels.}

\item{text}{Show correlation values as text.}

\item{pstar}{Show significance stars.}

\item{power}{Adjacency power exponent.}

\item{setMargins}{Adjust plot margins automatically.}

\item{mar1}{Margins for dendrogram panel.}

\item{mar2}{Margins for heatmap panel.}

\item{cex.lab}{Label character expansion factor.}

\item{cex.text}{Text character expansion factor.}

\item{plotDendro}{Plot the dendrogram panel.}

\item{plotHeatmap}{Plot the heatmap panel.}

\item{dendro.horiz}{Plot dendrogram horizontally.}

\item{dendro.width}{Relative width of dendrogram.}

\item{dendro.labels}{Show dendrogram labels.}

\item{nmax}{Maximum number of eigengenes shown.}

\item{fixclust}{Fix cluster ordering.}

\item{mask.intra}{Mask intra-module correlations.}

\item{justdata}{Return data without plotting.}
}
\value{
Correlation matrix (invisible), or matrix if justdata is TRUE.
}
\description{
Plot the adjacency correlation heatmap matrix of eigengenes with or
without traits. This can show how traits cluster together with the
eigengenes.
}
