% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotModuleScores}
\alias{plotModuleScores}
\title{Plot top modules most correlated with trait for multi expression
data.}
\usage{
plotModuleScores(
  res,
  trait,
  multi = FALSE,
  nmax = 16,
  collapse.trait = FALSE,
  plotlib = "base",
  setpar = TRUE
)
}
\description{
Plot top modules most correlated with trait for multi expression
data.
}
