% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dendro.R
\name{plotMultiDendroAndColors}
\alias{plotMultiDendroAndColors}
\title{Plot dendrograms for multiple WGCNA objects}
\usage{
plotMultiDendroAndColors(
  wgcna,
  block = 1,
  extra.colors = NULL,
  show.kme = FALSE,
  show.traits = FALSE,
  show.contrasts = FALSE,
  show.mat = NULL,
  clust = TRUE,
  use.tree = 0,
  rm.na = TRUE,
  sd.wt = 0,
  nmax = -1,
  main = "Dendro and colors",
  colorHeight = 0.5,
  marAll = c(0.4, 5, 1, 0.2),
  cex = 1
)
}
\arguments{
\item{block}{Block number to plot.}

\item{extra.colors}{Additional color rows to display.}

\item{show.kme}{Show module eigengene correlations.}

\item{show.traits}{Show trait correlations.}

\item{show.contrasts}{Show contrast correlations.}

\item{show.mat}{Custom matrix to show.}

\item{clust}{Cluster columns by correlation.}

\item{use.tree}{Tree index for consensus.}

\item{rm.na}{Remove NA-only columns.}

\item{sd.wt}{Standard deviation weighting factor.}

\item{nmax}{Maximum number of columns shown.}

\item{main}{Plot title string.}

\item{colorHeight}{Relative height of color rows.}

\item{marAll}{Margin sizes vector.}

\item{cex}{Character expansion factor.}

\item{multi_wgcna}{Named list of WGCNA objects.}
}
\value{
NULL (invisible). Generates a plot.
}
\description{
Plot dendrograms for multiple WGCNA objects
}
