% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rho2bluered}
\alias{rho2bluered}
\title{Convert correlation to blue-red colors}
\usage{
rho2bluered(R, a = 1, f = 0.95)
}
\arguments{
\item{R}{Numeric correlation matrix or vector.}

\item{a}{Exponent for nonlinear scaling.}

\item{f}{Color attenuation factor.}
}
\value{
Character matrix or vector of colors.
}
\description{
Converts correlation values [-1;1] to blue-white-red colors. Good
for creating color labels for labeledHeatmaps that expect colors.
NOTE: use WGCNA::numbers2colors???
}
\keyword{internal}
